+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The qhun-transpiler project uses [Semantic Versioning](https://semver.org/spec/v
 
 - Fixed a bug in lua and wow target that causes string literals in an object context to be generated without the nessesary computed property wrapper brackets.
 - Fixed a bug in wow target where path building for external files targeted the wrong file.
+- Fixed a bug in lua and wow target where MappedTypes with array values were not detected as arrays
 
 ##  **0.5.0** released 2018-10-20